removing non-template stuff. will use fully templated types from here on
