+++ include/xvstruct/internals.h
@@ -12,7 +12,9 @@ typedef uint8_t pbuf_type;
 
 namespace internals {
 
-// iteration function for generating maximum value
+////////////////////////////////////////////////////////////////////////////////////////////////////////
+/// iteration function for generating maximum value
+////////////////////////////////////////////////////////////////////////////////////////////////////////
 template <typename T, uint16_t Sz>
 struct MaskMax {
   enum : T {
@@ -103,7 +105,9 @@ struct Packer<double, Sz> {
 };
 
 
-// this iteration is from forArg ... 1; 0 is special case to stop
+////////////////////////////////////////////////////////////////////////////////////////////////////////
+/// this iteration is from forArg ... 1; 0 is special case to stop
+////////////////////////////////////////////////////////////////////////////////////////////////////////
 template <typename T, uint16_t B, uint16_t forArg, uint16_t offset>
 struct ForLoop {
   static T LEGet(pbuf_type* buffer, T x) {
@@ -172,7 +176,10 @@ void RawIF<T, pb, Sz>::setLE(pbuf_type* buffer, T x) {
   }
 }
 
-// helper function to find next matching bit alignment
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////
+/// helper function to find next matching bit alignment
+////////////////////////////////////////////////////////////////////////////////////////////////////////
 template<uint16_t b, uint16_t AlignBit>
 struct nextAlignedBit {
   enum : uint16_t{
@@ -226,7 +233,9 @@ struct IndexedLE <T, b, Sz, 0> {
   }
 };
 
-// Temporary Object used by array
+////////////////////////////////////////////////////////////////////////////////////////////////////////
+/// Temporary Object used by array
+////////////////////////////////////////////////////////////////////////////////////////////////////////
 template<typename T, uint16_t b, uint16_t Sz, uint16_t N>
 struct LEArrayTemp {
   enum : uint16_t { N_minus_one = N - 1 };
@@ -276,12 +285,10 @@ struct BoolArrayTemp {
   }
 };
 
+
+
 }  // namespace internals
-
-
 }  // namespace xvstruct
-
-
